move from destinations to destination
